change bottom bar to bottom sheet
