+++ lib/src/views/payment/payment_view.dart
@@ -4,6 +4,7 @@ import 'package:flutter_loyalty_point/src/styles/styles.dart';
 import 'package:flutter_loyalty_point/src/utils/types/category_product_type.dart';
 import 'package:flutter_loyalty_point/src/view_models/payment/payment_view_model.dart';
 import 'package:flutter_loyalty_point/src/views/payment/widgets/payment_product.dart';
+import 'package:flutter_loyalty_point/src/views/widgets/bottom_sheet_widget.dart';
 import 'package:heroicons/heroicons.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
@@ -11,7 +12,7 @@ import 'package:provider/provider.dart';
 part 'widgets/payment_app_bar.dart';
 part 'widgets/payment_method_list.dart';
 part 'widgets/payment_product_card.dart';
-part 'widgets/payment_bottom_bar.dart';
+part 'widgets/payment_bottom_sheet.dart';
 
 class PaymentView extends StatelessWidget {
   const PaymentView({super.key});
@@ -28,7 +29,7 @@ class PaymentView extends StatelessWidget {
           PaymentMethodList(),
         ],
       ),
-      bottomNavigationBar: PaymentBottomBar(),
+      bottomNavigationBar: PaymentBottomSheet(),
     );
   }
 }
