+++ lib/src/views/product_list/product_list_view.dart
@@ -4,6 +4,7 @@ import 'package:flutter_loyalty_point/src/styles/styles.dart';
 import 'package:flutter_loyalty_point/src/utils/types/category_product_type.dart';
 import 'package:flutter_loyalty_point/src/utils/types/view_state_type.dart';
 import 'package:flutter_loyalty_point/src/view_models/product_list/product_list_view_model.dart';
+import 'package:flutter_loyalty_point/src/views/widgets/bottom_sheet_widget.dart';
 import 'package:heroicons/heroicons.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
@@ -12,7 +13,7 @@ part 'widgets/product_list_app_bar.dart';
 part 'widgets/product_list_form_field.dart';
 part 'widgets/product_list_information_card.dart';
 part 'widgets/product_list_product_list.dart';
-part 'widgets/product_list_bottom_bar.dart';
+part 'widgets/product_list_bottom_sheet.dart';
 part 'widgets/product_list_product_card.dart';
 
 class ProductListView extends StatelessWidget {
@@ -42,7 +43,7 @@ class ProductListView extends StatelessWidget {
         ),
 
         // bottom bar section
-        bottomNavigationBar: ProductListBottomBar(),
+        bottomSheet: ProductListBottomSheet(),
       ),
     );
   }
